Add form sign up to subscribe widget view
